requirements.txt
tensorflow
numpy
pandas
pillow
streamlit
matplotlib
keras
opencv-python-headless
